Enable building in MSYS2 environment

+++ build.sh
@@ -3,6 +3,8 @@
 # Define how to build the libraries and executables:
 BUILD_TYPE=Debug
 Fortran_COMPILER=gfortran
+GENERATOR="GNU Makefiles"
+ENABLE_OPI_SUPPORT=OFF
 LIBSUFFIX="so"
 if [[ "$OSTYPE" == "linux-gnu" ]]; then
   LIBSUFFIX="so"
@@ -12,6 +14,9 @@ elif [[ "$OSTYPE" == "CYGWIN"* ]]; then
   LIBSUFFIX="dll"
 elif [[ "$OSTYPE" == "MINGW"* ]]; then
   LIBSUFFIX="dll"
+elif [[ "$OSTYPE" == "msys"* ]]; then
+  LIBSUFFIX="dll"
+  GENERATOR="Ninja"
 fi
 git submodule update --init --recursive
 ################################################################################
@@ -19,25 +24,28 @@ git submodule update --init --recursive
 #                                  Build OPI                                   #
 #                                                                              #
 ################################################################################
-# Build OPI
-cd OPI || exit
-# Create the build directory if it does not exist
-if [[ ! -d "build" ]]; then
-  mkdir build
-else
-  rm -rf build/*
+if [[ "$ENABLE_OPI_SUPPORT" == "ON" ]]; then
+  # Build OPI
+  cd OPI || exit
+  # Create the build directory if it does not exist
+  if [[ ! -d "build" ]]; then
+    mkdir build
+  else
+    rm -rf build/*
+  fi
+  cd build || exit
+  echo "Updating cmake"
+  cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_INSTALL_PREFIX=../ -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_CXX11=ON -DENABLE_CL_SUPPORT=OFF -DENABLE_CUDA_SUPPORT=OFF -DENABLE_PYTHON=OFF -G $GENERATOR ../
+  echo "Building OPI"
+  cmake --build .
+  cmake --install .
+  if [[ $? -ne 0 ]]; then
+    echo "Could not build OPI. Exiting."
+    exit $?
+  fi
+  echo "Leaving OPI"
+  cd ../../
 fi
-cd build || exit
-echo "Updating cmake"
-cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_INSTALL_PREFIX=../ -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_CXX11=ON -DENABLE_CL_SUPPORT=OFF -DENABLE_CUDA_SUPPORT=OFF -DENABLE_PYTHON=OFF ../
-echo "Building OPI"
-make install
-if [[ $? -ne 0 ]]; then
-  echo "Could not build OPI. Exiting."
-  exit $?
-fi
-echo "Leaving OPI"
-cd ../../
 ################################################################################
 #                                                                              #
 #                                Build libslam                                 #
@@ -53,9 +61,10 @@ else
 fi
 cd build || exit
 echo "Updating cmake"
-cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OpenMP_SUPPORT=OFF -DENABLE_POSTGRESQL_SUPPORT=OFF ../
+cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OpenMP_SUPPORT=OFF -DENABLE_POSTGRESQL_SUPPORT=OFF -G $GENERATOR ../
 echo "Building libslam"
-make install
+cmake --build .
+cmake --install .
 if [[ $? -ne 0 ]]; then
     echo "Could not build libslam. Exiting."
     exit $?
@@ -71,23 +80,23 @@ cd ../
 #                                Build pFUnit                                  #
 #                                                                              #
 ################################################################################
-cd pFUnit || exit
-# Create the build directory if it does not exist
-if [[ ! -d "build" ]]; then
-  mkdir build
-else
-  rm -rf build/*
-fi
-cd build || exit
-echo "Updating cmake"
-export FC=$Fortran_COMPILER
-cmake -DSKIP_MPI=yes ../
-echo "Building pFUnit"
-make
-make install
-cd ../
-echo "Leaving pFUnit"
-cd ../
+# cd pFUnit || exit
+# # Create the build directory if it does not exist
+# if [[ ! -d "build" ]]; then
+#   mkdir build
+# else
+#   rm -rf build/*
+# fi
+# cd build || exit
+# echo "Updating cmake"
+# export FC=$Fortran_COMPILER
+# cmake -DSKIP_MPI=yes -G $GENERATOR ../
+# echo "Building pFUnit"
+# cmake --build .
+# cmake --install .
+# cd ../
+# echo "Leaving pFUnit"
+# cd ../
 ################################################################################
 #                                                                              #
 #                                 Build NEPTUNE                                #
@@ -97,11 +106,13 @@ cd ../
 if [[ ! -d "lib" ]]; then
   mkdir lib
 fi
-# Create the links to libraries needed by CAMP
+# Create the links to libraries needed by NEPTUNE
 cd lib || exit
 ln -sf ../libslam/lib/libslam-Fortran.$LIBSUFFIX .
-ln -sf ../OPI/lib/libOPI-Fortran.$LIBSUFFIX .
-ln -sf ../OPI/lib/libOPI.$LIBSUFFIX .
+if [[ "$ENABLE_OPI_SUPPORT" == "ON" ]]; then
+  ln -sf ../OPI/lib/libOPI-Fortran.$LIBSUFFIX .
+  ln -sf ../OPI/lib/libOPI.$LIBSUFFIX .
+fi
 cd ..
 # Create the include directory if it does not exist
 if [[ ! -d "include" ]]; then
@@ -110,7 +121,9 @@ fi
 # Create the links to includes needed by NEPTUNE
 cd include || exit
 ln -sf ../libslam/include/SLAM .
-ln -sf ../OPI/include/OPI .
+if [[ "$ENABLE_OPI_SUPPORT" == "ON" ]]; then
+  ln -sf ../OPI/include/OPI .
+fi
 cd ..
 # Create the build directory if it does not exist
 if [[ ! -d "build" ]]; then
@@ -121,9 +134,10 @@ fi
 cd build || exit
 echo "Updating cmake"
 export PFUNIT_DIR=..//pFUnit/build/installed
-cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OPI_SUPPORT=ON -DSKIP_MSIS_2=ON ../
+cmake -DCMAKE_BUILD_TYPE=$BUILD_TYPE -DCMAKE_Fortran_COMPILER=$Fortran_COMPILER -DENABLE_OPI_SUPPORT=$ENABLE_OPI_SUPPORT -DSKIP_MSIS_2=ON -G $GENERATOR ../
 echo "Building NEPTUNE"
-make install
+cmake --build .
+cmake --install .
 if [[ $? -ne 0 ]]; then
     echo "Could not build NEPTUNE. Exiting."
     exit $?
